Fixed sword buying.
The entire time I was trying to remove from the database the currently bought sword whilst the one that need to be removed was the one the user had before he buys a sword. Stupid but took awhile to notice. There were also small changes to the entity classes. I also tried to do some changes to the Battlefield and Battlefield monster classes but that went nowhere and was  scrapped.