Update to admin functions + logout
Added some extra admin functions, plus made it so that on logout user data gets saved and in case he is "fighting" a monster it gets sent back to its battlefield. Also made it so extra swords are added to the shop if the amount in it is too low. Have to change that later to happen more smoothly.